Fix repeated audio plays on mobile
This was a rather tricky problem because the default solution using
HTML5 audio elements works just fine on desktop, but, for some reason,
mobile platforms throttle and impose some weird restrictions on playing
audio. To be specific, the issue is that repeated "clicks" or taps on
mobile would not trigger repeated audio plays the sound would "lock up"
and feel janky. Also, without `preventDefault` repeated taps (or double
clicks, I guess) would zoom in on the button element.

This is the result of some googling and iterative refactoring using
ChatGPT. I've added some explanation of what's going on in a comment
block in the script section, but essentially this solution revolves
around using the more powerful Web Audio API and keeping track of the
source that was playing in between taps.

Kinda yucky when this is something expected to work ootb.